refactor: how sample data is generated so both unit tests and cypress tests can use the same process
