+++ cypress/tsconfig.json
@@ -11,7 +11,8 @@
     "resolveJsonModule": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["../../*"]
+      "@/*": ["../*"],
+      "@test-utils/*": ["../test-utils/*"]
     }
   },
   "include": [
