+++ test/models/customer.spec.js
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import Chance from 'chance';
 import { CustomerModel } from '../../application/api/models/customer.ts';
 import * as databaseService from '../../application/api/services/databaseService';
-import * as testDataGenerator from '../testDataGenerator';
+import * as testDataGenerator from '../../test-utils/testDataGenerator.ts';
 
 const chance = Chance();
