+++ test/models/finish.spec.js
@@ -3,7 +3,7 @@ const chance = Chance();
 import mongoose from 'mongoose';
 import { FinishModel } from '../../application/api/models/finish.ts';
 import * as databaseService from '../../application/api/services/databaseService';
-import * as testDataGenerator from '../testDataGenerator';
+import * as testDataGenerator from '../../test-utils/testDataGenerator.ts';
 
 describe('validation', () => {
     let finishAttributes;
