+++ test/services/baseProductService.spec.js
@@ -1,6 +1,6 @@
 import { BaseProductModel } from '../../application/api/models/baseProduct.ts';
 import * as databaseService from '../../application/api/services/databaseService';
-import * as testDataGenerator from '../testDataGenerator';
+import * as testDataGenerator from '../../test-utils/testDataGenerator.ts';
 import { FinishModel } from '../../application/api/models/finish.ts';
 import { MaterialModel } from '../../application/api/models/material.ts';
 import { CustomerModel } from '../../application/api/models/customer.ts';
